Migrated notifications tests to webpack
