+++ corehq/apps/notifications/static/notifications/spec/bootstrap3/main.js
@@ -1,0 +1,8 @@
+import hqMocha from "mocha/js/main";
+import "commcarehq";
+
+import "notifications/spec/bootstrap3/service_spec";
+
+$(function () {
+    hqMocha.run();
+});
