+++ corehq/apps/notifications/static/notifications/spec/bootstrap5/main.js
@@ -1,0 +1,8 @@
+import hqMocha from "mocha/js/main";
+import "commcarehq";
+
+import "notifications/spec/bootstrap5/service_spec";
+
+$(function () {
+    hqMocha.run();
+});
