+++ corehq/apps/notifications/templates/notifications/spec/bootstrap5/mocha.html
@@ -1,16 +1,10 @@
 {% extends 'mocha/base.html' %}
 {% load hq_shared_tags %}
 
-{% block dependencies %}
-  <script src="{% static 'notifications/js/bootstrap5/notifications_service.js' %}"></script>
-{% endblock %}
+{% js_entry "notifications/spec/bootstrap5/main" %}
 
 {% block fixtures %}
   <ul>
     {% include 'notifications/partials/bootstrap5/global_notifications.html' %}
   </ul>
-{% endblock %}
-
-{% block mocha_tests %}
-  <script src="{% static 'notifications/spec/bootstrap5/NotificationsService.spec.mo.js' %}"></script>
 {% endblock %}
